domain/src/main/java/com/example/gomesrodris/archburgers/domain/services/CardapioServices.java
package com.example.gomesrodris.archburgers.domain.services;

import com.example.gomesrodris.archburgers.domain.entities.ItemCardapio;
import com.example.gomesrodris.archburgers.domain.repositories.ItemCardapioRepository;
import com.example.gomesrodris.archburgers.domain.valueobjects.TipoItemCardapio;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CardapioServices {
    private final ItemCardapioRepository itemCardapioRepository;

    @Autowired
    public CardapioServices(ItemCardapioRepository itemCardapioRepository) {
        this.itemCardapioRepository = itemCardapioRepository;
    }

    public List<ItemCardapio> listLanches() {
        return itemCardapioRepository.findByTipo(TipoItemCardapio.LANCHE);
    }

}
